Skip inappropriate test cases when running as root
