+++ timberjack_linux_test.go
@@ -4,15 +4,59 @@ package timberjack
 
 import (
 	"os"
+	"path/filepath"
 	"syscall"
 	"testing"
 )
 
-func TestOpenNewDefaultPerm(t *testing.T) {
-	// if runtime.GOOS == "windows" {
-	// 	t.Skip("Skipping default perm test on Windows")
-	// }
+func TestRotate_OpenNewFails(t *testing.T) {
+	if os.Getuid() == 0 {
+		t.Skip("Skipping test when running as root")
+	}
+	badPath := "/bad/path/logfile.log"
+	if _, err := os.Stat(badPath); err == nil {
+		t.Skip("Skipping test that relies on non-existent path")
+	}
+	l := &Logger{
+		Filename: badPath,
+	}
+	// force an invalid path to trigger openNew failure
+	err := l.rotate("manual")
+	if err == nil {
+		t.Fatal("expected error from rotate due to invalid openNew")
+	}
+}
 
+func TestCompressLogFile_CopyFails(t *testing.T) {
+	if os.Getuid() == 0 {
+		t.Skip("Skipping test when running as root")
+	}
+	dir := t.TempDir()
+	src := filepath.Join(dir, "bad.log")
+	dst := src + ".gz"
+
+	if err := os.WriteFile(src, []byte("data"), 0o200); err != nil { // write-only
+		t.Fatalf("failed to create test file: %v", err)
+	}
+	defer os.Chmod(src, 0o644)
+
+	originalStat := osStat
+	osStat = func(name string) (os.FileInfo, error) {
+		return os.Stat(src)
+	}
+	defer func() { osStat = originalStat }()
+
+	l := &Logger{}
+	// snapshot patched osStat
+	l.resolveConfigLocked()
+
+	err := l.compressLogFile(src, dst)
+	if err == nil {
+		t.Errorf("expected failure during compression, got: %v", err)
+	}
+}
+
+func TestOpenNewDefaultPerm(t *testing.T) {
 	// Ensure no bits get masked out.
 	syscall.Umask(0o000)
 
@@ -30,10 +74,6 @@ func TestOpenNewDefaultPerm(t *testing.T) {
 }
 
 func TestOpenNewCustomPerm(t *testing.T) {
-	// if runtime.GOOS == "windows" {
-	// 	t.Skip("Skipping custom perm test on Windows")
-	// }
-
 	// Ensure no bits get masked out.
 	syscall.Umask(0o000)
